+++ cloisim_ros_bringup/package.xml
@@ -17,6 +17,7 @@
 
   <depend>launch_ros</depend>
   <depend>rclcpp</depend>
+  <depend>cloisim_ros_bringup_param</depend>
   <depend>cloisim_ros_websocket_service</depend>
   <depend>cloisim_ros_micom</depend>
   <depend>cloisim_ros_lidar</depend>
