+++ cloisim_ros_bringup_param/README.md
@@ -1,0 +1,3 @@
+# cloisim_ros_bringup_param
+
+The `cloisim_ros_bringup_param` package is an bringup parameters for cloisim_ros.
